RESOURCES/WEB/RESSOURCES PYTHON/DIVERS/ad_library_watch/requirements.txt
streamlit>=1.31
requests>=2.31
python-dateutil>=2.8
beautifulsoup4>=4.12
playwright>=1.40
